Ceres/codebase/api-django/requirements.txt
ansible==2.9.2
boto3==1.17.85
botocore==1.20.85
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
cryptography==3.4.7
Django==2.2
django-cors-headers==3.7.0
django-environ==0.4.5
django-storages==1.11.1
djangorestframework==3.11.0
firebase-admin==4.4.0
gcloud==0.17.0
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.6.0
grpcio==1.38.0
gunicorn==20.0.4
httplib2==0.15.0
idna==2.10
Jinja2==3.0.1
jmespath==0.10.0
jws==0.1.3
MarkupSafe==2.0.1
msgpack==1.0.0
oauth2client==3.0.0
oauthlib==3.1.0
Pillow==7.0.0
protobuf==3.12.4
psycopg2==2.8.5
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
PyJWT==2.0.0
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2019.3
PyYAML==5.4.1
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.0
s3transfer==0.4.2
sentry-sdk==0.14.4
six==1.13.0
sqlparse==0.3.0
twilio==6.50.1
uritemplate==3.0.1
urllib3==1.25.9
